+++ lib/hi_energy_ai/resources/deeplinks.rb
@@ -3,8 +3,9 @@
 module HiEnergyAi
   module Resources
     class Deeplinks < Resource
-      def generate(attributes, **params)
-        post("/deeplinks/generate", params: params, body: attributes)
+      def generate(attributes = UNSET, **params)
+        body_attrs, query = split_attributes(attributes, params)
+        post("/deeplinks/generate", params: query, body: body_attrs)
       end
     end
   end